Change the 'workout' property on Lift to be optional
[#116942929]

+++ WorkoutTracker/Workout/FileSystem/LiftDeleteAgent.swift
@@ -8,13 +8,15 @@ public class LiftDeleteAgent {
     }
     
     public func delete(lift: Lift) {
-        let liftNameSnakeCased = lift.name.stringByReplacingOccurrencesOfString(" ", withString: "_")
-        let fileName = "Lifts/\(liftNameSnakeCased)/\(String(lift.workout.timestamp)).json"
-        
-        do {
-            try localStorageWorker.deleteFile(fileName)
-        } catch {
-            print("Failed to delete file with name \(fileName) (for lift named \"\(lift.name)\"")
+        if let liftWorkout = lift.workout {
+            let liftNameSnakeCased = lift.name.stringByReplacingOccurrencesOfString(" ", withString: "_")
+            let fileName = "Lifts/\(liftNameSnakeCased)/\(String(liftWorkout.timestamp)).json"
+            
+            do {
+                try localStorageWorker.deleteFile(fileName)
+            } catch {
+                print("Failed to delete file with name \(fileName) (for lift named \"\(lift.name)\"")
+            }
         }
     }
 }
